wip: collapsing endpoints

+++ src/controllers/cached-realms.controller.ts
@@ -3,13 +3,9 @@ import { Controller, HttpCode, HttpStatus, Inject, UnprocessableEntityException 
 import { ApiTags } from '@nestjs/swagger';
 import { Cache } from 'cache-manager';
 
-import { DeleteConfigIds, DeleteRealm, GetRealm } from '@/decorators/controller.method.decorators';
+import { DeleteRealm, GetRealm } from '@/decorators/controller.method.decorators';
 import { ParamRealm, QueryConfigIds, QueryRealm } from '@/decorators/controller.parameter.decorators';
-import {
-  OpenApi_DeleteRealm,
-  OpenApi_DeleteRealmConfigIds,
-  OpenApi_GetRealm,
-} from '@/decorators/open-api.controller.decorators';
+import { OpenApi_DeleteRealm, OpenApi_GetRealm } from '@/decorators/open-api.controller.decorators';
 import { reduceToConfigs } from '@/helpers/reduce-to-configs.helper';
 import { ConfigFactoryService } from '@/services/config-factory.service';
 import { RealmsService } from '@/services/realms.service';
@@ -25,7 +21,7 @@ export class CachedRealmsController {
 
   @GetRealm()
   @OpenApi_GetRealm()
-  async getRealmConfigIds(@QueryRealm() realm: string, @QueryConfigIds() configIds?: string[]) {
+  async getRealm(@QueryRealm() realm: string, @QueryConfigIds() configIds?: string[]) {
     const postfix = `$${realm} @${this.configFactory.config.namespacePostfix}`;
     let cache = (await this.cache.get(postfix)) ?? ({} as any);
 
@@ -50,21 +46,19 @@ export class CachedRealmsController {
   @DeleteRealm()
   @OpenApi_DeleteRealm()
   @HttpCode(HttpStatus.NO_CONTENT)
-  async deleteRealm(@ParamRealm() realm: string) {
-    await this.cache.del(`${realm}_${this.configFactory.config.namespacePostfix}`);
-    await this.realmsService.deleteRealm(realm);
-  }
-
-  @DeleteConfigIds()
-  @OpenApi_DeleteRealmConfigIds()
-  @HttpCode(HttpStatus.NO_CONTENT)
-  async deleteRealmConfigIds(@ParamRealm() realm: string, @QueryConfigIds() configIds: string[]) {
+  async deleteRealm(@ParamRealm() realm: string, @QueryConfigIds() configIds?: string[]) {
     const postfix = `$${realm} @${this.configFactory.config.namespacePostfix}`;
+
+    if (!configIds) {
+      await this.cache.del(`${realm}_${this.configFactory.config.namespacePostfix}`);
+      return await this.realmsService.deleteRealm(realm);
+    }
+
     const ids = Array.from(new Set(configIds.filter((e) => e)));
     const cache = (await this.cache.get(postfix)) ?? ({} as any);
     const keys = Object.keys(cache).filter((key) => delete cache[configIds.find((id) => id === key)]);
     await this.realmsService.deleteRealmConfigIds(realm, ids);
     if (keys.length) await this.cache.set(postfix, cache, this.configFactory.config.ttl);
-    else await this.cache.del(postfix);
+    else return await this.cache.del(postfix);
   }
 }
